+++ src/main/java/com/github/microprograms/micro_oss_core/Transaction.java
@@ -6,11 +6,11 @@ import com.github.microprograms.micro_oss_core.model.dml.InsertCommand;
 import com.github.microprograms.micro_oss_core.model.dml.UpdateCommand;
 
 public interface Transaction extends AutoCloseable {
-    int deleteObject(DeleteCommand command) throws MicroOssTransactionException;
+    long deleteObject(DeleteCommand command) throws MicroOssTransactionException;
 
-    int insertObject(InsertCommand command) throws MicroOssTransactionException;
+    long insertObject(InsertCommand command) throws MicroOssTransactionException;
 
-    int updateObject(UpdateCommand command) throws MicroOssTransactionException;
+    long updateObject(UpdateCommand command) throws MicroOssTransactionException;
 
     void commit() throws MicroOssTransactionException;
